requirements.txt
Django>=5.1,<6.0
matplotlib>=3.9,<4.0
numpy>=2.1,<3.0
oracledb>=2.4,<3.0
python-dotenv>=1.0,<2.0
